WIP: refactoring tasks into modules
